Use Maven build tool
This enables downloading the jar dependencies instead of bundling them
within the repository.
