+++ .gitignore
@@ -1,0 +1,2 @@
+/c3m.jar
+/target/
