--- c3m-code/src/log4j.properties
@@ -1,6 +1,0 @@
-log4j.rootLogger=ERROR, stdout
-log4j.logger.c3m=INFO
-log4j.logger.c3m.constraint=DEBUG
-log4j.appender.stdout=org.apache.log4j.ConsoleAppender
-log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
-log4j.appender.stdout.layout.ConversionPattern=%d [%-5p] (%F:%M:%L) %m%n
